# swagger.yaml
openapi: 3.0.0
info:
  title: Nagar Parishad School Leaving Certificate API
  version: 1.0.0
  description: RESTful API for managing Nagar Parishad School Leaving Certificates with audit trail support
  contact:
    name: API Support
    email: support@nagarparishad.gov.in
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.kaamlo.com
    description: Production server
...
paths:
...
  /api/schools:
    get:
      summary: Get all schools
      tags: [Schools]
      responses:
        '200':
          description: List of all schools
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/School'
        '500':
          $ref: '#/components/responses/ServerError'
    
    post:
      summary: Create a new school
      tags: [Schools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - district
              properties:
                name:
                  type: string
                  example: स्व. जतिरामजी बर्वे नगर परिषद प्राथमिक शाळा, रामटेक
                district:
                  type: string
                  example: नागपूर
                state:
                  type: string
                  example: महाराष्ट्र
                created_by:
                  type: integer
                  description: User ID who created this record
      responses:
        '201':
          description: School created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/School'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/schools/{id}:
    get:
      summary: Get school by ID
      tags: [Schools]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: School ID
      responses:
        '200':
          description: School details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/School'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      summary: Update school
      tags: [Schools]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/School'
      responses:
        '200':
          description: School updated successfully
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete school
      tags: [Schools]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: School deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/students:
    get:
      summary: Get all students
      tags: [Students]
      responses:
        '200':
          description: List of all students
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
    
    post:
      summary: Create a new student
      tags: [Students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - full_name
                - date_of_birth
              properties:
                full_name:
                  type: string
                  example: राजेश कुमार
                date_of_birth:
                  type: string
                  format: date
                  example: '2010-05-15'
                created_by:
                  type: integer
      responses:
        '201':
          description: Student created successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Student ID or Aadhar already exists
  /api/students/{id}:
    get:
      summary: Get student by ID
      tags: [Students]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student details
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      summary: Update student
      tags: [Students]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Student updated successfully
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete student
      tags: [Students]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /api/students/search/{identifier}:
    get:
      summary: Search student by Student ID or Aadhar number
      tags: [Students]
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
          description: Student ID or Aadhar number
      responses:
        '200':
          description: Student found
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/auth/otp/verify
  schemas:
...
    School:
      type: object
      required: [name, district]
      properties:
        id:
          type: integer
          description: School ID
        name:
          type: string
          description: School name
          example: स्व. जतिरामजी बर्वे नगर परिषद प्राथमिक शाळा, रामटेक
        address:
          type: string
          description: School address
        taluka:
          type: string
          example: रामटेक
        district:
          type: string
          example: नागपूर
        state:
          type: string
          example: महाराष्ट्र
        phone_no:
          type: string
          example: '07112234567'
        email:
          type: string
          format: email
        general_register_no:
          type: string
          example: GR001
        school_recognition_no:
          type: string
        udise_no:
          type: string
        affiliation_no:
          type: string
        board:
          type: string
          default: Maharashtra State
        medium:
          type: string
          default: Marathi
        status:
          type: string
          enum: [active, inactive, archived]
          default: active
        created_by:
          type: integer
          description: User ID who created the record
        updated_by:
          type: integer
          description: User ID who last updated the record
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Student:
      type: object
      required: [full_name, date_of_birth]
      properties:
        id:
          type: integer
        student_id:
          type: string
          example: STU001
        uid_aadhar_no:
          type: string
          pattern: '^[0-9]{12}$'
          description: 12-digit Aadhar number
          example: '123456789012'
        full_name:
          type: string
          example: राजेश कुमार
        father_name:
          type: string
          example: राम कुमार
        mother_name:
          type: string
          example: गीता देवी
        surname:
          type: string
          example: शर्मा
        nationality:
          type: string
          example: Indian
        mother_tongue:
          type: string
          example: Marathi
        religion:
          type: string
        caste:
          type: string
        sub_caste:
          type: string
        birth_place_village:
          type: string
          example: रामटेक
        birth_place_taluka:
          type: string
        birth_place_district:
          type: string
          example: नागपूर
        birth_place_state:
          type: string
          example: महाराष्ट्र
        birth_place_country:
          type: string
          default: India
        date_of_birth:
          type: string
          format: date
          example: '2010-05-15'
        date_of_birth_words:
          type: string
          example: पंधरा मे दोन हजार दहा
        status:
          type: string
          enum: [active, inactive, archived]
          default: active
        created_by:
          type: integer
        updated_by:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: object
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            errors:
              - msg: Field is required
                param: field_name
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Internal server error